+++ lib/widgets.php
@@ -12,6 +12,7 @@
  * @return void
  */
 function calendar_ui_widgets_init() {
-    
+    // add user or site calendar
+    elgg_register_widget_type('calendar', elgg_echo('calendar_ui:calendar:widget'), elgg_echo('calendar_ui:calendar:widget:description'), array("profile", "dashboard"), false);
     
 }