+++ views/default/widgets/calendar/edit.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * Elgg Calendar UI plugin
+ * @package calendar_ui
+ */
+
+
+
+
+
+
+