+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## 2.2.0 (2016-09-25)
+### 2.1.1 (2016-10-26)
+- Added calendar widget
+- Added calendar menu item on Owner's menu
+
+### 2.1.0 (2016-09-25)
 * **releases:** Initial release